Move contract and signature to vm
1. Move contract and signature to vm
