+++ api/common/interfaces.go
@@ -7,16 +7,16 @@ import (
 
 	"github.com/gogo/protobuf/proto"
 	. "github.com/nknorg/nkn/block"
-	"github.com/nknorg/nkn/ledger"
 	"github.com/nknorg/nkn/common"
-	"github.com/nknorg/nkn/contract"
 	"github.com/nknorg/nkn/errors"
+	"github.com/nknorg/nkn/ledger"
 	"github.com/nknorg/nkn/node"
 	"github.com/nknorg/nkn/pb"
 	. "github.com/nknorg/nkn/transaction"
 	"github.com/nknorg/nkn/util/address"
 	"github.com/nknorg/nkn/util/config"
 	"github.com/nknorg/nkn/util/log"
+	"github.com/nknorg/nkn/vm/contract"
 )
 
 const (
