+++ ledger/blockValidator.go
@@ -13,10 +13,10 @@ import (
 	. "github.com/nknorg/nkn/errors"
 	. "github.com/nknorg/nkn/pb"
 	"github.com/nknorg/nkn/por"
-	"github.com/nknorg/nkn/signature"
 	. "github.com/nknorg/nkn/transaction"
 	"github.com/nknorg/nkn/util/config"
 	"github.com/nknorg/nkn/util/log"
+	"github.com/nknorg/nkn/vm/signature"
 )
 
 const (
