+++ ledger/mining.go
@@ -9,10 +9,10 @@ import (
 	"github.com/nknorg/nkn/crypto/util"
 	. "github.com/nknorg/nkn/pb"
 	"github.com/nknorg/nkn/por"
-	"github.com/nknorg/nkn/signature"
 	. "github.com/nknorg/nkn/transaction"
 	"github.com/nknorg/nkn/util/config"
 	"github.com/nknorg/nkn/vault"
+	"github.com/nknorg/nkn/vm/signature"
 	"github.com/nknorg/nnet/log"
 )
