+++ vm/contract/ContractType.go
