+++ vm/contract/contract.go
