+++ vm/contract/contractBuilder.go
