+++ vm/contract/contractParameter.go
