+++ vm/contract/pubkeyParameter.go
