+++ vm/signature/signer.go
