+++ vm/signature/validation.go
